don't interpret escape sequences in pattern definitions
Previously, patterns like the GREEDYMULTILINE in the tests would cause the
config parse to fail because the '\n' was rendered as a new line. Rather
than guessing what the translation from acceptable PCREe escapes is (most
like a safe thing to do, but you never know), just don't interpret the
escapes at all.